chall 48/Projet48hpython/requirements.txt
aniso8601==8.1.1
attrs==20.3.0
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.10.0
pymongo==3.11.3
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1
